+++ milkman-jdbc/src/main/java/milkman/plugin/jdbc/AbstractJdbcProcessor.java
@@ -1,12 +1,15 @@
 package milkman.plugin.jdbc;
 
+import milkman.plugin.jdbc.domain.RowSetResponseAspect;
+import org.apache.commons.io.IOUtils;
+
+import java.io.IOException;
+import java.sql.Blob;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.util.LinkedList;
 import java.util.List;
-
-import milkman.plugin.jdbc.domain.RowSetResponseAspect;
 
 public class AbstractJdbcProcessor {
 
@@ -20,12 +23,25 @@ public class AbstractJdbcProcessor {
 		rowSetAspect.setColumnNames(columnNames);
 		
 		while(resultSet.next() && !maxRowLimitReached(rowSetAspect)) {
-			List<Object> row = new LinkedList<Object>();
+			List<String> row = new LinkedList<>();
 			for(int i = 1; i <= metaData.getColumnCount(); ++i) {// column idx starts at 1
-				row.add(resultSet.getObject(i));
+				var value = resultSet.getObject(i);
+				row.add(valueToString(value));
 			}
 			rowSetAspect.addRow(row);
 		}
+	}
+
+	private String valueToString(Object value) {
+		if (value instanceof Blob) {
+			try {
+				value = IOUtils.toString(((Blob) value).getBinaryStream());
+			} catch (IOException | SQLException e) {
+				value = "BLOB";
+			}
+		}
+		String stringValue = value != null ? value.toString() : "NULL";
+		return stringValue;
 	}
 
 	private boolean maxRowLimitReached(RowSetResponseAspect rowSetAspect) {