+++ milkman-jdbc/src/main/java/milkman/plugin/jdbc/JdbcRequestPlugin.java
@@ -1,10 +1,5 @@
 package milkman.plugin.jdbc;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.concurrent.CountDownLatch;
-
-import javafx.application.Platform;
 import lombok.SneakyThrows;
 import milkman.domain.RequestContainer;
 import milkman.domain.RequestExecutionContext;
@@ -14,15 +9,10 @@ import milkman.plugin.jdbc.domain.JdbcSqlAspect;
 import milkman.plugin.jdbc.editor.JdbcRequestEditor;
 import milkman.plugin.jdbc.editor.JdbcResultSetAspectEditor;
 import milkman.plugin.jdbc.editor.SqlAspectEditor;
-import milkman.ui.main.dialogs.StringInputDialog;
-import milkman.ui.plugin.CustomCommand;
-import milkman.ui.plugin.RequestAspectEditor;
-import milkman.ui.plugin.RequestAspectsPlugin;
-import milkman.ui.plugin.RequestTypeEditor;
-import milkman.ui.plugin.RequestTypePlugin;
-import milkman.ui.plugin.ResponseAspectEditor;
-import milkman.ui.plugin.Templater;
-import milkman.utils.AsyncResponseControl.AsyncControl;
+import milkman.ui.plugin.*;
+
+import java.util.Collections;
+import java.util.List;
 
 public class JdbcRequestPlugin implements RequestTypePlugin, RequestAspectsPlugin {