+++ milkman-jdbc/src/main/java/milkman/plugin/jdbc/domain/RowSetResponseAspect.java
@@ -1,16 +1,16 @@
 package milkman.plugin.jdbc.domain;
 
-import java.util.LinkedList;
-import java.util.List;
-
 import lombok.Data;
 import milkman.domain.ResponseAspect;
+
+import java.util.LinkedList;
+import java.util.List;
 
 @Data
 public class RowSetResponseAspect implements ResponseAspect {
 
-	List<String> columnNames = new LinkedList<String>();
-	List<List<Object>> rows = new LinkedList<List<Object>>();
+	List<String> columnNames = new LinkedList<>();
+	List<List<String>> rows = new LinkedList<>();
 	
 	
 	@Override
@@ -19,7 +19,7 @@ public class RowSetResponseAspect implements ResponseAspect {
 	}
 
 
-	public void addRow(List<Object> row) {
+	public void addRow(List<String> row) {
 		rows.add(row);
 	}