+++ milkman-jdbc/src/main/java/milkman/plugin/jdbc/domain/TableResponseContainer.java
@@ -2,7 +2,7 @@ package milkman.plugin.jdbc.domain;
 
 import milkman.domain.ResponseContainer;
 
-public class JdbcResponseContainer extends ResponseContainer {
+public class TableResponseContainer extends ResponseContainer {